Add files via upload
The design of the private event on Federal Hill eloquently embodied the theme of "A New Era for the World and Opportunities for Malaysia." With four elegantly set round tables accommodating 24 guests each, along with six cocktail tables for more informal gatherings, the layout provided an inviting and versatile space for networking and discussions. The registration table ensured a seamless check-in process, while the dedicated area for prayer offered a respectful and inclusive environment. The centerpiece of the design was a warm and cozy zone, carefully curated for open communication and impactful speeches, creating an atmosphere conducive to meaningful dialogues and the exchange of ideas.